Providing a login shell in ssh connections
The `exec_command` method of the paramiko module sends a command through a
non-login, non-interactive shell. Such a shell has a meager environment,
since the `/etc/profile` standard file is never sourced. This generates
errors in many of the `WfxTestDut` methods such as the `read_*_version`
family, which calls `wfx_test_agent` without knowing where it's located.
It also leaves the class user with no other option than to manually pass
environment variables to the `run` method.

Ensure that a login shell is passed on every command with the properly
escaped string.