Convert project to a python package (library)
Add `__init__.py` files to the top directory and subpackage folders to
convert the project into a packageable python library. Modify the import
paths accordingly. Since python package names should not contain dashes,
rename the existing folders with underscores.

NOTE: ideally, the whole project should be packaged and uploaded to the
PyPi repository for external projects to import it as a dependency. This
would require the addition of a `pyproject.toml` and a pipeline that
builds and publishes the package accordingly.