Source pds definition file from package resources
The PDS definition file used as a template for configuring the WFX chip
is currently sourced from a hardcoded path. This path is not valid when
the library is imported from an outside project and has to be manually
renamed. This leads to several compatibility problems and is considered
bad practice. Instead, source the file through the `importlib.resources`
module provided by the standard library. This ensures that the file will
be transparently sourced regardless of the library installation path.